Move test files into a separate folder
- To be more consistent with our other projects. It also
  allows some of the test-only rules for ESLint to work
  properly
