+++ test/services/SensorApiService.test.js
@@ -1,27 +1,27 @@
 // In-house dependencies
-import ButtonsApiService from './ButtonsApiService'
-import * as FetchService from './FetchService'
+import SensorApiService from '../../src/services/SensorApiService'
+import * as FetchService from '../../src/services/FetchService'
 
 // Setup mocks for this whole test file
-jest.mock('./FetchService', () => ({
+jest.mock('../../src/services/FetchService', () => ({
   post: jest.fn(),
 }))
-jest.mock('./CredentialsService', () => ({
+jest.mock('../../src/services/CredentialsService', () => ({
   getApiKey: jest.fn().mockReturnValue('API_KEY'),
 }))
 
-describe('ButtonsApiService', () => {
+describe('SensorApiService', () => {
   describe('testRequest', () => {
     it('calls POST with the correct URI, base URL, and headers', () => {
       const expectedParameters = {
         uri: '/alert/test',
-        base: 'https://buttons.com',
+        base: 'https://sensor.com',
         headers: {
           'X-API-KEY': 'API_KEY',
         },
       }
 
-      ButtonsApiService.testRequest()
+      SensorApiService.testRequest()
 
       expect(FetchService.post).toHaveBeenCalledWith(expectedParameters)
     })
@@ -31,13 +31,13 @@ describe('ButtonsApiService', () => {
     it('calls POST with the correct URI, base URL, and headers', () => {
       const expectedParameters = {
         uri: 'not/in/real/life',
-        base: 'https://buttons.com',
+        base: 'https://sensor.com',
         headers: {
           'X-API-KEY': 'API_KEY',
         },
       }
 
-      ButtonsApiService.fakeEndpointRequest()
+      SensorApiService.fakeEndpointRequest()
 
       expect(FetchService.post).toHaveBeenCalledWith(expectedParameters)
     })
